third_party/WebKit/LayoutTests/fast/dom/tabindex-defaults-expected.txt

Default value of tabIndex IDL attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS anchor.tabIndex is 0
PASS button.tabIndex is 0
PASS input.tabIndex is 0
PASS select.tabIndex is 0
PASS textarea.tabIndex is 0
FAIL editableDiv.tabIndex should be 0. Was -1.
PASS normalDiv.tabIndex is -1
PASS anchor.setAttribute("tabindex", "invalid"); anchor.tabIndex is 0
PASS button.setAttribute("tabindex", "invalid"); button.tabIndex is 0
PASS input.setAttribute("tabindex", "invalid"); input.tabIndex is 0
PASS select.setAttribute("tabindex", "invalid"); select.tabIndex is 0
PASS textarea.setAttribute("tabindex", "invalid"); textarea.tabIndex is 0
FAIL editableDiv.setAttribute("tabindex", "invalid"); editableDiv.tabIndex should be 0. Was -1.
PASS normalDiv.setAttribute("tabindex", "invalid"); normalDiv.tabIndex is -1
PASS anchor.setAttribute("tabindex", "9999999999"); anchor.tabIndex is 0
PASS button.setAttribute("tabindex", "9999999999"); button.tabIndex is 0
PASS input.setAttribute("tabindex", "9999999999"); input.tabIndex is 0
PASS select.setAttribute("tabindex", "9999999999"); select.tabIndex is 0
PASS textarea.setAttribute("tabindex", "9999999999"); textarea.tabIndex is 0
FAIL editableDiv.setAttribute("tabindex", "9999999999"); editableDiv.tabIndex should be 0. Was -1.
PASS normalDiv.setAttribute("tabindex", "9999999999"); normalDiv.tabIndex is -1
PASS successfullyParsed is true

TEST COMPLETE
